Clean up aspect ratio mode
