+++ packages/themes/aspect-4-3.js
@@ -1,0 +1,6 @@
+import aspect from './aspect'
+
+export default {
+  ...aspect,
+  aspectRatio: 9 / 16,
+}
